requirements.txt
yfinance
pandas
numpy
python-telegram-bot
flask
pytz
requests
gspread
google-auth
python-dotenv